+++ pyscriptjs/tests/integration/test_py_config.py
@@ -270,7 +270,6 @@ class TestConfig(PyScriptTest):
             </py-script>
             """
         )
-        assert self.console.log.lines[0] == self.PY_COMPLETE
         assert self.console.log.lines[-2:] == [
             "hello from A",
             "hello from B",
@@ -313,5 +312,4 @@ class TestConfig(PyScriptTest):
             </py-script>
             """
         )
-        assert self.console.log.lines[0] == self.PY_COMPLETE
         assert self.console.log.lines[-1] == "hello from A"